Split query and delete tests to their own files
Makes it more understandable what tests concern what functionality.
Needed to split a few tests apart as they were testing both query and
delete functionality.
